Client only & deprecation fix
Basically, updated code to new forge standarts:
Menu mobs is client side only mod, no need to use it on server, thus
removed proxys and marked as clientSideOnly. Without this, if mod is
installed on server, forge will think that it must be installed on
clients, while it is actually not compulsory.
FMLCommonHandler.instance().bus() and MinecraftForge.EVENT_BUS are same
thing now.
