--- src/main/java/superbas11/MenuMobs/CommonProxy.java
@@ -1,6 +1,0 @@
-package superbas11.MenuMobs;
-
-public class CommonProxy {
-    protected void registerMainMenuTickHandler() {
-    }
-}
